+++ src/main/java/me/zilzu/mycoupon/storage/CouponEntity.java
@@ -14,10 +14,10 @@ public class CouponEntity {
     public final CouponCurrency couponCurrency;
     public final DiscountType discountType;
     public final Long amountOff;
-    public final Float percentOff;
+    public final Double percentOff;
     public final LocalDateTime createdTime;
 
-    public CouponEntity(String id, CouponDuration duration, Integer durationInMonth, CouponCurrency couponCurrency, DiscountType discountType, Long amountOff, Float percentOff, LocalDateTime createdTime) {
+    public CouponEntity(String id, CouponDuration duration, Integer durationInMonth, CouponCurrency couponCurrency, DiscountType discountType, Long amountOff, Double percentOff, LocalDateTime createdTime) {
         this.id = id;
         this.duration = duration;
         this.durationInMonth = durationInMonth;
